proglanglearn/requirements.txt
autopep8==1.4.4
backports.csv==1.0.7
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-ckeditor==5.7.1
django-debug-toolbar==1.11
django-htmlmin==0.11.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-recaptcha3==0.1.0
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
jsmin==2.2.2
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
